Deleting attachments when weapons/armor deleted (#531)
In this PR
Iterates over attachments when armor or weapons are deleted on an actor sheet
_onAlterationDelete() is a public method (used by actor-sheet.mjs), so I'm removing the _

Testing
Optional steps for some background
On beta, populate an actor sheet with a weapon and give it effects and upgrades
Delete the weapon
Do game.actors.get('<actor ID>').items. The weapon will be gone, but the weaponEffects and upgrades remain
Follow the steps above on this branch. The weapon, weaponEffects, and upgrades will now be gone
Do the same for armor