Added Default Faction to Role (#858)
In this PR
Added an Automatic Faction id drop to roles
Added drop code to attach the faction to the actor if they don't already have a faction
Testing
add a role with a default faction and validate that both the role and faction and any attached perks are all added.
add a role to an actor with a faction already and validate only the role and attached perks are added.