Additional perk updates (#876)
In this PR
Added movement increase as an option
Added perk Advancement
Testing
Create a perk select Has Choice and select movement and an increase. Make sure the character has some sort of base speed from an Origin.
Create a perk and select Has Advancement. Pick some settings, and drop that perk on to a character, and then drop the perk a second time. Character should have the perk once, but the name should change.